:adhesive_bandage: (sonarcloud): Fix bugs and code smells
:adhesive_bandage: (otm8009a): Use std::byte for bitwise operation
:adhesive_bandage: (jpeg): Remove redundant cast
:adhesive_bandage: (dma2d): Fix implicit conversion loses
:adhesive_bandage: (jpeg): Add tests recommanded by sonarcloud
:adhesive_bandage: (hal): Remove unused acces specifier
:adhesive_bandage: (hal): Use =default as default implementation
:adhesive_bandage: const non-mutable member function
:adhesive_bandage: (ll): Remove unused namespace
:adhesive_bandage: (dsi): Avoid implicit conversion
:adhesive_bandage: (lcd) Add missing final keyword
:adhesive_bandage: (font): Add const in fontGetPixelBytes argument
:adhesive_bandage: (otm8009a): Replace "/*" by "//"
:adhesive_bandage: (otm8009a): Nested namespace

:adhesive_bandage: (sonarcloud): Fix implicit conversion loses integer precision

https://sonarcloud.io/project/issues?id=leka_LekaOS&issues=AXiE6yzYQJJ2tBYqGH4h&open=AXiE6yzYQJJ2tBYqGH4h&pullRequest=140

:adhesive_bandage: (sonarcloud): Fix Use "=default" instead of the default implementation of this special member functions

https://sonarcloud.io/project/issues?id=leka_LekaOS&issues=AXiE6y3oQJJ2tBYqGH5j&open=AXiE6y3oQJJ2tBYqGH5j&pullRequest=140

:adhesive_bandage: (sonarcloud): Fix Add the "explicit" keyword to this constructor

:adhesive_bandage: (sonarcloud): Explicitly use reinterpret_cast instead of C-style cast

:adhesive_bandage: (sonarcloud): Fix Merge this "if" statement with the enclosing one

https://sonarcloud.io/project/issues?id=leka_LekaOS&issues=AXiE6ywSQJJ2tBYqGH2k&open=AXiE6ywSQJJ2tBYqGH2k&pullRequest=140

:adhesive_bandage: (sonarcloud): Fix field '_dsi' will be initialized after field '_backlight'

:adhesive_bandage: (sonarcloud): Define command array as const

:adhesive_bandage: (sonarcloud): Fix duplicated blocks of code must be removed

:adhesive_bandage: (sonarcloud): Fix Concatenate this namespace with the nested one

:adhesive_bandage: (sonarcloud): Fix Remove the misleading "/*" characters

:adhesive_bandage: (sonarcloud): Fix Rename this namespace to match the regular expression

+++ drivers/LKCoreVideo/include/internal/corevideo_config.h
@@ -41,7 +41,7 @@ namespace dsi {
 
 	constexpr uint32_t lcdClock			 = 27429;	// LcdClk = 27429 kHz
 	constexpr uint32_t laneByteClock_kHz = 62500;	// 500 MHz / 8 = 62.5 MHz = 62500 kHz
-	constexpr float txEscapeClockDiv	 = laneByteClock_kHz / 15620;
+	constexpr uint32_t txEscapeClockDiv	 = laneByteClock_kHz / 15620;
 
 }	// namespace dsi
